auth/requirements.txt
astroid==3.2.2
blinker==1.8.2
click==8.1.7
dill==0.3.8
Flask==3.0.3
Flask-MySQLdb==2.0.0
importlib_metadata==7.1.0
isort==5.13.2
itsdangerous==2.2.0
jedi==0.19.1
Jinja2==3.1.4
MarkupSafe==2.1.5
mccabe==0.7.0
mysqlclient==2.2.4
parso==0.8.4
platformdirs==4.2.2
PyJWT==2.8.0
pylint==3.2.3
tomli==2.0.1
tomlkit==0.12.5
typing_extensions==4.12.2
Werkzeug==3.0.3
zipp==3.19.2
